Decouple LastRunMetrics class from Llama and reuse it for Mistral
